Move public facing classes in online-banking-core to internal package
These classes shouldn't be public facing, since they are abstract.

COAND-724
