+++ online-banking-core/src/main/java/com/adyen/checkout/onlinebankingcore/internal/OnlineBankingComponent.kt
@@ -5,7 +5,7 @@
  *
  * Created by atef on 20/9/2022.
  */
-package com.adyen.checkout.onlinebankingcore
+package com.adyen.checkout.onlinebankingcore.internal
 
 import androidx.annotation.RestrictTo
 import androidx.lifecycle.LifecycleOwner
