+++ online-banking-sk/src/main/java/com/adyen/checkout/onlinebankingsk/OnlineBankingSKConfiguration.kt
@@ -12,7 +12,7 @@ import android.content.Context
 import com.adyen.checkout.action.GenericActionConfiguration
 import com.adyen.checkout.components.core.Amount
 import com.adyen.checkout.core.Environment
-import com.adyen.checkout.onlinebankingcore.OnlineBankingConfiguration
+import com.adyen.checkout.onlinebankingcore.internal.OnlineBankingConfiguration
 import kotlinx.parcelize.Parcelize
 import java.util.Locale
